// frontend/src/generated/ledger-api.d.ts
export interface components {
    schemas: {
// ...
        Command: {
            CreateAndExerciseCommand: components["schemas"]["CreateAndExerciseCommand"];
        } | {
            CreateCommand: components["schemas"]["CreateCommand"];
        } | {
            ExerciseByKeyCommand: components["schemas"]["ExerciseByKeyCommand"];
        } | {
            ExerciseCommand: components["schemas"]["ExerciseCommand"];
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}